feat的功能是增加了选择JSON文件的功能，解析JSON数据并更新Excel文件中的车位状态。具体改动如下：- 在ControlPanel中添加了选择文件按钮 - 在ParkingAnnotator中实现了选择文件的逻辑，包括读取和解析JSON文件- 更新了ExcelHandler以支持更新车位状态的功能 - 在Constants中添加了新的错误消息常量- 修改了HousewindowList中的presaleid字段类型

+++ src/main/java/com/yuda/parking/HousewindowList.java
@@ -25,7 +25,7 @@ public class HousewindowList {
     private double otherarea;
     private long permitid;
     private String permitno;
-    private long presaleid;
+    private int presaleid;
     private long price;
     private long prjid;
     private String projectname;
